fix checkstyle

+++ src/main/java/ru/yandex/practicum/filmorate/dal/FilmRepository.java
@@ -98,7 +98,7 @@ public class FilmRepository extends BaseRepository<Film> implements FilmStorage 
     }
 
     @Override
-    public void deleteFilm(Integer filmId){
+    public void deleteFilm(long filmId) {
         delete(DELETE_FILM_QUERY, filmId);
     }
